+++ NEWS.md
@@ -1,3 +1,5 @@
 ## glitchr 0.0.1.9000
 ---------------------
+* New `progress` function tracks `squaregram` progress
+* New `squaregram` function makes square a directory of images
 * New `square` function makes square images